We won 4 out of 5 times with my new offensive agent.
Offensive agent runs away from ghosts when they get close, and eats pacman if he's on the way to a dot.

Some improvements that can be made:
use distance to our side as a feature when running away
avoid the place you last died, or prefer the opposite side of the board from where you died.

git-svn-id: svn+ssh://njoubert.com/var/svn@783 997b1fe7-5dcb-cc6a-e049-b7da040a1fbe