My initial attempt.
This uses a producer-consumer model in the ChatState, but it is a bit stupid because it forces only one thread at a time to be accessing the state. Really you want to allow as many consumers as they want to read the state, and limit it to only one producer at a time (with no comsumers reading while the producer is there).

The ChatServer really just spawns a new thread for every request coming in.

git-svn-id: svn+ssh://njoubert.com/var/svn@1192 997b1fe7-5dcb-cc6a-e049-b7da040a1fbe